requirements.txt

Django==4.2.5
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
Pillow==10.0.0
psycopg2-binary==2.9.7
virtualenv==20.24.5

